# Optimize IntData & ScopeData