--- a/asdar/.editorconfig
+++ b/asdar/.editorconfig
@@ -1,0 +1,9 @@
+[*]
+end_of_line = lf
+charset = utf-8
+insert_final_newline = true
+indent_size = 4
+
+[*.{c,h}]
+indent_style = tab
+max_line_length = off