--- a/asdar/src/objects/bool.h
+++ b/asdar/src/objects/bool.h
@@ -13,4 +13,9 @@ Object *boolobj_c2asda(bool cbool);
 // asserts that the object is boolobj_true or boolobj_false
 bool boolobj_asda2c(Object *asdabool);
 
+/** Negate a Boolean
+ * Returns a new reference
+ */
+Object *boolobj_neg(Object *obj);
+
 #endif   // OBJECTS_BOOL_H