… asdar/src/partialfunc.c
… asdar/tests/objects/test_func.c