Thought this could help performance, and it seems to do so quite well. 

A small benchmark I ran showed ~2x speedup due to this! :confetti_ball: 

I ran a small `ripgrep` and noticed `IntData->mpz` is only accessed from `src/objects/int.c`, so this commit *shouldn't* break anything + `make test` passes.

Here's the benchmark I ran:

```shell
❯ cat fizzbuzz.asda
# assumes that x and y are positive, and something else that i'm not sure about
let divisible = (Int x, Int y) -> Bool:
    let x_copy = x    # this is a bug in asdac, this shouldn't be required

    #       x divisible by y
    # <=>   there exists an integer n so that x = n*y
    # <=>   if x = n*y + r, where 0 <= r < y, then r=0
    #
    # trying to find the r
    while TRUE:
        for let r = 0; r != y; r = r+1:
            if x_copy == y + r:
                return (r == 0)

        # exists n so that x = n*y  <=>  exists m so that x-y = m*y
        x_copy = x_copy - y

let and = (Bool a, Bool b) -> Bool:
    if a:
        if b:
            return TRUE
    return FALSE

let fizzbuzz = (Int n) -> Str:
    if (n `divisible` 3) `and` (n `divisible` 5):
        return "Fizzbuzz"
    if n `divisible` 3:
        return "Fizz"
    if n `divisible` 5:
        return "Buzz"
    return n.to_string()

# loop starts at 5 because my stupid divisible func breaks otherwise
for let n = 5; n != 4096; n = n+1:
    print(fizzbuzz(n))

❯ git checkout cf641f1
❯ make -j8 NDEBUG=true
❯ make run FILE=fizzbuzz.asda > /dev/null

real    0m8,166s
user    0m8,166s
sys     0m0,000s

❯ git checkout 6338e1a
❯ make -j8 NDEBUG=true
❯ make run FILE=fizzbuzz.asda > /dev/null

real    0m4,193s
user    0m4,175s
sys     0m0,000s
```

## Update
After the ScopeData commit, performance is increased oncemore!

```shell
❯ git checkout 0fef16a
❯ make run NDEBUG=true FILE=fizzbuzz.asda  > /dev/null

real    0m3,696s
user    0m3,676s
sys     0m0,004s
```

This PR now is a ~54.73% speedup! Amazing!